refactor: reorganize platforms code further
